find referring sites
Google Analytics  > Filter Manager > Add Filter

Custom filters:
Exclude: This type of filter excludes log file lines (hits) that match the Filter Pattern. Matching lines are ignored in their entirety: for example, a filter that excludes Chrome also excludes all other information in that log line, such as visitor, path, referral, and domain information.
Include: This type of filter includes log file lines (hits) that match the Filter Pattern. All non-matching hits are ignored and any data in non-matching hits is unavailable in the reports.
Lowercase / Uppercase: Converts the contents of the field into all uppercase or all lowercase characters. These filters only affect letters, and do not affect special characters or numbers.
Search & Replace: This is a simple filter that you can use to search for a pattern within a field and replace the found pattern with an alternate form.
Advanced: This type of filter allows you to build a field from one or two other fields. The filtering engine applies the expressions in the two Extract fields to the specified fields and then constructs a third field using the Constructor expression.